requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
requests==2.31.0
urllib3==2.0.2
Werkzeug==2.3.4
Flask-SQLAlchemy
Flask-Migrate
redis
Pillow
python-dotenv
psycopg2-binary
